pyproject.toml
[project]
name = "agents-playground"
version = "0.1.0"
description = "A playground for experimenting with AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "google-generativeai>=0.8.5",
    "requests>=2.32.4",
    "lxml>=5.4.0",
    "pygithub>=2.6.1",
    "markdown>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents_playground"]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
